services/ai-tutor/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.6
langchain-core==0.2.10
langchain-community==0.2.6
langgraph==0.1.19
langchain-groq==0.1.6
qdrant-client==1.10.0
sentence-transformers==3.0.1
sqlalchemy==2.0.31
asyncpg==0.29.0
aioredis==2.0.1
pydantic==2.8.0
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
